+++ src/Ledger/RolloverLedger/RolloverLedgerView/index.js
@@ -1,2 +1,1 @@
-export { default } from './RolloverLedgerContainer';
 export { RolloverLedgerView } from './RolloverLedgerView';