+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Show alert when the transfer will result in a negative amount. Refs UIF-427.
 * *BREAKING*: Update `@folio/stripes` to `8.0.0`. Refs UIF-430.
 * Add rollover budget "Cash balance" setting to rollover screen. Refs UIF-440.
+* Add possibility to display FY rollover settings for each rollover AND test rollover. Refs UIF-443.
 
 ## [3.3.2](https://github.com/folio-org/ui-finance/tree/v3.3.2) (2022-11-30)
 [Full Changelog](https://github.com/folio-org/ui-finance/compare/v3.3.1...v3.3.2)